Refactor keystore configuration files creation logic to KeystoreFileManager
