+++ core/src/test/java/tech/pegasys/web3signer/core/service/http/handlers/keymanager/delete/KeystoreFileManagerTest.java
@@ -10,10 +10,10 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.core.service.http.handlers.keymanager.imports.util;
+package tech.pegasys.web3signer.core.service.http.handlers.keymanager.delete;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static tech.pegasys.web3signer.core.service.http.handlers.keymanager.imports.util.KeystoreConfigurationFilesCreator.YAML_MAPPER;
+import static tech.pegasys.web3signer.core.multikey.metadata.parser.YamlSignerParser.YAML_MAPPER;
 
 import tech.pegasys.web3signer.core.multikey.metadata.FileKeyStoreMetadata;
 import tech.pegasys.web3signer.core.multikey.metadata.SigningMetadata;
@@ -25,12 +25,11 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-class KeystoreConfigurationFilesCreatorTest {
-
+class KeystoreFileManagerTest {
   @Test
   void configurationFilesAreCreated(@TempDir Path parentDir) throws Exception {
-    new KeystoreConfigurationFilesCreator(parentDir, "filename", "{\"test\":true}")
-        .createFiles("password".toCharArray());
+    new KeystoreFileManager(parentDir)
+        .createKeystoreFiles("filename", "{\"test\":true}", "password".toCharArray());
 
     final Path metadataYamlFile = parentDir.resolve("filename.yaml");
     final Path keystoreJsonFile = parentDir.resolve("filename.json");
@@ -43,8 +42,8 @@ class KeystoreConfigurationFilesCreatorTest {
 
   @Test
   void yamlContentIsValidFileKeyStoreMetadata(@TempDir Path parentDir) throws Exception {
-    new KeystoreConfigurationFilesCreator(parentDir, "filename", "{\"test\":true}")
-        .createFiles("password".toCharArray());
+    new KeystoreFileManager(parentDir)
+        .createKeystoreFiles("filename", "{\"test\":true}", "password".toCharArray());
 
     final Path metadataYamlFile = parentDir.resolve("filename.yaml");
 
@@ -57,8 +56,8 @@ class KeystoreConfigurationFilesCreatorTest {
 
   @Test
   void yamlContentIsNotConverted(@TempDir Path parentDir) throws Exception {
-    new KeystoreConfigurationFilesCreator(parentDir, "filename", "{\"test\":true}")
-        .createFiles("password".toCharArray());
+    new KeystoreFileManager(parentDir)
+        .createKeystoreFiles("filename", "{\"test\":true}", "password".toCharArray());
 
     final Path metadataYamlFile = parentDir.resolve("filename.yaml");
     final Path keystoreJsonFile = parentDir.resolve("filename.json");
@@ -77,8 +76,8 @@ class KeystoreConfigurationFilesCreatorTest {
 
   @Test
   void passwordContentsAreWritten(@TempDir Path parentDir) throws Exception {
-    new KeystoreConfigurationFilesCreator(parentDir, "filename", "{\"test\":true}")
-        .createFiles("password".toCharArray());
+    new KeystoreFileManager(parentDir)
+        .createKeystoreFiles("filename", "{\"test\":true}", "password".toCharArray());
 
     final Path keystorePasswordFile = parentDir.resolve("filename.password");
 
@@ -87,8 +86,8 @@ class KeystoreConfigurationFilesCreatorTest {
 
   @Test
   void jsonDataIsWritten(@TempDir Path parentDir) throws Exception {
-    new KeystoreConfigurationFilesCreator(parentDir, "filename", "{\"test\":true}")
-        .createFiles("password".toCharArray());
+    new KeystoreFileManager(parentDir)
+        .createKeystoreFiles("filename", "{\"test\":true}", "password".toCharArray());
 
     final Path keystoreJsonFile = parentDir.resolve("filename.json");
